refactor: move shared.ts to fixtures/resources.ts
Media resource constants now live at fixtures/resources.ts, shared
across all framework apps. A thin re-export at apps/vite/src/resources.ts
keeps generated page imports clean.
