+++ apps/e2e/apps/vite/src/resources.ts
@@ -1,0 +1,2 @@
+/** Re-export media resources from the shared fixtures. */
+export { MEDIA } from '../../../fixtures/resources';
